Batch inference on the CLI
Adds --file and --output-dir arguments to cog infer.
Doesn't add batch infer capabilities to the Python library.
